+++ home/moz4r/0_INSTALL.bat
@@ -1,0 +1,4 @@
+taskkill.exe /F /IM java.exe
+taskkill.exe /F /IM javaW.exe
+taskkill.exe /F /IM chrome.exe
+java -jar myrobotlab.jar -invoke python execFile %cd%/INMOOV-AI_install.py -service GUIService GUIService python Python
