+++ home/moz4r/1_START_THE_ROBOT.bat
@@ -1,0 +1,4 @@
+taskkill.exe /F /IM java.exe
+taskkill.exe /F /IM javaW.exe
+taskkill.exe /F /IM chrome.exe
+java -jar myrobotlab.jar -invoke python execFile %cd%/INMOOV-AI_startup.py -service GUIService GUIService python Python
