+++ home/moz4r/INMOOV-AI_install.py
@@ -1,0 +1,1 @@
+Runtime.install()
