+++ home/moz4r/INMOOV-AI_move_head_random.py
@@ -1,10 +1,13 @@
 MoveHeadTimer = Runtime.start("MoveHeadTimer","Clock")
-MoveHeadTimer.setInterval(1000)
+MoveHeadTimer.setInterval(10)
 
 
 def MoveHead(timedata):
-	if IsInmoovLeft==1:
-		i01.setHeadSpeed(0.8, 0.8)
+	if IsInmoovArduino==1:
+		i01.setHeadSpeed(0.1, 0.1)
 		i01.moveHead(random.randint(50,130),random.randint(50,130))
+		HeadSide.moveTo(random.randint(50,130))
+		HeadSide.setSpeed(0.1)
+		MoveHeadTimer.setInterval(1000)
 	
 MoveHeadTimer.addListener("pulse", python.name, "MoveHead")
