+++ home/moz4r/Inmoov_AI/AIML/FRENCH/phrases_aleatoires.aiml
@@ -72,9 +72,9 @@ je t écoute , quel est mon nom ?
 				<li>#SHORTBREATH01# Salut, pourquoi m'avoir réveillé ?</li>
 				<li>A vos ordres, humain !</li>
 				<li>A vos ordres, mammifère !</li>
-				<li>Humanoide à ton service !</li>
-				<li>Humanoide à ton écoute !</li>
-				<li>Oui !! Oui !! suis là ? je t'écoute  !</li>
+				<li>Humanohide à ton service !</li>
+				<li>Humanohide à ton écoute !</li>
+				<li>ouioui ! je suis là ! je t'écoute  !</li>
 				<li>Tu crois que tu peux m'éteindre et m'allummer comme ça à ta convenance ?</li>
 				<li>Bonjour, que puis-je faire pour toi</li>
 				<li>Coucou ! Tu sais, je n'étais pas vraiment endormi</li>
