+++ home/moz4r/Inmoov_AI/AIML/FRENCH/son1.aiml
@@ -7,7 +7,7 @@
 			<mrl>
 				<service>python</service>
 				<method>exec</method>
-				<param>mouth.audioFile.playFileBlocking("..\sound\je_suis_ton_pere.mp3")</param>
+				<param>mouth.audioFile.playFileBlocking("sound\je_suis_ton_pere.mp3")</param>
 			</mrl>
       </oob></template>
 </category>
@@ -16,9 +16,11 @@
 			<mrl>
 				<service>python</service>
 				<method>exec</method>
-				<param>mouth.audioFile.playFileBlocking("..\sound\R2D2.mp3")</param>
+				<param>mouth.audioFile.playFileBlocking("sound\R2D2.mp3")</param>
 			</mrl>
       </oob></template>
+</category>
+<category><pattern>QUI EST R2D2</pattern><template><srai>QUI EST R2 D2</srai></template>
 </category>
 <category><pattern># STAR WARS #</pattern>
 <template><random>
