+++ home/moz4r/Inmoov_AI/BDD/birthday.csv
@@ -1,0 +1,4 @@
+02/1/1920,Isaac Asimov
+23/06/1912,Alan Turing
+4/09/1927,John McCarthy
+6/09/1980,John Doe
